Added oldid parameter to any requests that might happen
Now you can parse old versions of a page. Note that this is only
working with GET requests, but I couldn't find a place where the
TemplateRequest constructor was called in a POST. If there is,
basically the solution is the same except it should use req.body
and not req.query.

Change-Id: I25e457f45a3197074cde8ad9a49e3c75ef8610a4