Prefix all relative links with ./
This makes sure that we don't have to escape colons, which can otherwise
fool browsers into treating parts of titles as protocols. We could
alternatively add this prefix only for titles which actually contain colons to
save some chars.

In any case stripping ./ or (../)+ will yield the full link target.

Change-Id: I063614593be4bac7b58affda21ce161650911e27