Remove chaining capability to side-step subtype issues
It would be great to return a reference to the instance from void methods to enable
chaining, but this is difficult to combine with subclassing since the return
type of inherited methods will normally be that of the base class. A solution
to this can likely be found, but for now it seems to be easier to simply
disable chaining and make those methods return void instead. Can look into
this later if this bothers us too much.

Change-Id: I457ad5e68fc707cf11b762cbe48254794058b373