Fix 'src is undefined' error in the web service
subbu added this line in the wikitext->HTML web endpoint, and I
have no idea why. src doesn't exist in this function, or at least
it doesn't anymore. This will fix parsing of normal wikitext for
now, but may break what subbu tried to do.

gwicke: env.text is needed for the retrieval of original wikitext based on tsr
(for rt info). Set it from the proper variable in the wikitext -> html form
mode.

Change-Id: I99cdf6e016df33d6cdf15a8273760825cd1ad75c