handle shell errors in runtests.sh
Whenever node has an error, the stderr is intercepted to a result file
and nothing is shown. That might confuse a newcomer. This change
intercept the node exit code and will show the content of the result
file for user to investigate.

A typical use case is a node module depency missing.

Change-Id: I946771aea4289e2a5389dbca8b928d83722b234a