Fix bug from the Util-shuffle that breaks the API
Pretty major oversight from our changes in the getParserEnv function,
which is now in Util: We couldn't require the localsettings.js file
because it's in the js/api/ directory, not the js/lib/ one. Now we do
the include in js/api/ParserService.js again, so the crisis is averted.

Change-Id: I8c0b0cec057d907a5ed9261fcda6c2ffb6dded79