Fix a few OOM errors
Async returns with ret.tokens being a string were processed character-wise,
which quickly led to out-of-error conditions.

For now we work around this by wrapping the returned string in an array and
pass it on, but eventually we should investigate the emitters of these
unexpected results.

The switch parser function in particular seemed to emit a chunk which was
forced through toString(), something to investigate later.

No changes in parserTests, but pages like [[:en:Serie A 2011-12]] are now
round-tripping.

Change-Id: I9231f35663313319a17dd7096a7bb0ec91f34a94