Don't serialize auto-closed quotes
Use information available in the quote handler to set the autoInsertedEnd flag
and suppress serialization of the end tag if that is set in the serializer.

This also avoids a class of semantic diffs we were producing when serializing
empty bold/italic elements:

''
would serialize to
''''
which then parses as
<b>'</b>

Greens six more parser rt tests.

Change-Id: I7b3bcad8a6d5056245997aec18030d9c76e08d62