Support rt of hrs followed by content on same line
On pages like [[List of Mile High Series One episodes]], horizontal rules are
followed by content on the same page. Our serializer so far always added a
newline after a hr, which broke round-tripping for these lines.

This patch adds a lineContent flag on the hr that signals the content on the
same line, and suppresses the newline after the hr if set. New horizontal
lines (without data-parsoid set) are still serialized on a line of their own.

No changes in parser tests: parserTestsNeeded!

Change-Id: Iaf8381a9ba39e3a37f6be6b6131c530b10736fdc