Bug 42178: Use scaled images
So far we linked to full-size images, which can be really large. The reason
for this is mainly that traditional MediaWiki image scaling can only be
accessed by calling into the API, using action=parse or (if it exists) a
dedicated thumbnail API.

The WMF cluster's url-based on-demand scaling thankfully allows us to get away
with a cheap URL-based hack in the meantime, which is what is implemented in
this patch.

No changes to parser tests, as all thumb-related tests are failing because of
the differing HTML output currently. I have however extended the HTML output
normalization to strip the size pre/postfixes, and verified that the only
output diffs are in the un-normalized HTML.

Change-Id: I876316c3e83118252d0ee0537e5db8d211be9562