Require a space or newline before a generic tag attribute
So far we did not require this, which caused tag-like constructs like <khyô> in
http://parsoid.wmflabs.org:8001/result/b91dc07cd0132da4b2daf3f0bc99a20912931130/Romanization%20of%20Bengali
to be parsed as <key ô=''>.

This patch fixes this by requiring a space or newline before each attribute in
a tag.

There are no changes in parserTests apart from a change in the tokenization of
the (already falling)
<span <includeonly>id="v1"</includeonly><noinclude>id="v2"</noinclude>>bar</span>.
A parser test for <khyô> is under way.

Change-Id: Iac4b5c480c5b2a43cf80741ae4ea3b7e1b963488