Make sure the serializer initialState remains unchanged
extendProps is just a shallow assignment of passed-in values, so for anything
that is not assigned by value (objects for example) it would just assign a
reference. These objects are later modified and will thus modify the
initial serializer state for later serializer runs.

This patch clones both options and initialState to avoid this issue. It also
freezes the initialState to make sure any modification is detected early. My
hope is that this will fix the random list issue we see in production.

There are no changes in parserTests.

Change-Id: Ic71e384b89241cf96bbff2186d2ea19945ed9c12