Round-trip treebuilder-stripped HTML end tags
Use the mw:EndTag shadow tag to detect stripped end tags, and insert an
mw:Placeholder for round-tripping. The source for the mw:Placeholder is not
yet tsr-based, would need to do so in the HTML5Treebuilder where we still have
access to the environment and the original tokens.

With this patch, this test case now round-trips fine:

<code>
 foo</code>

Three more parser tests are passing.

Change-Id: I6b9a730cc4735f43d5a70d061abea591fab73f46