Fix regression introduced by 6710479b6748e15053c8e1cacb64a4471b1ce1df
Newline-only template parameters were no longer matched after protecting the
sol production in block elements against inline breaks. This caused some
template transclusions to be parsed as plain text, which was then
nowiki-escaped in the serializer causing rt failures in testing. That code was
not deployed, so hooray for rt testing ;)

With this patch, some of the tests added in
https://gerrit.wikimedia.org/r/38750 are now passing (+5 green). One of the
two tests does not pass as the PHP parser emits <p><br/></p> for a simple
{{echo|\n}}, which looks more like a bug than a feature.

Change-Id: I9aa173ed8e90695dbd3ed94505f279b237853f77