First pass: normalize white-space in DOM.
This patch just retrieves leading/trailing nls in the DOM and
adds counts to the DSR array.  This patch won't help selser in
this form.  The code is more targeted at the regular serializer
assuming no access to source.  But, that assumption is no
longer useful and we are now going to whitespace source ranges
instead of nl counts and use original source in both selser and
regular serializer to output non-significant whitespace.

This patch needs to be reworked to track all ws rather than
just newlines.

Change-Id: Ibc8ae9073e70448295b397549438b4fb4470164a