Fix rt regression introduced by config fetch patch
The new environment created in roundtripDiff dropped the source on the floor,
which screwed up the WTS. We don't really need a new environment, so this
patch just passes the env in instead of the source.

Independently, the WTS is changed to directly use env.page.src instead of
custom src variables.

The use_source flag in parserTests is changed from defaulting to false to
true, which reflects the way we use the WTS in general now.

With this patch, wt2wt results improve from 703 to 789 green tests and selser
improves from 7369 to 8208. This is partly from the fixed regression and
partly from the use_source flag.

Change-Id: I86905c47c301bc47e48e2467eef2a1c08aa85c2b