Change wgScript to script in WikiConfig
To be more consistent with the existing naming scheme, I changed things around
in the WikiConfig object. Not a big change, and it doesn't affect much.

Also removed some unnecessary variables from the config.

Also fixed the debugging flags misstep that I committed.

Also fixed the parse.js script to use fetchTemplates and the preprocessor
when necessary or asked to.

Change-Id: Ia402c3cace4a3e6023d01a5065b1d9baca7d6fdf