Use pretty serialization for parse.js HTML output too
Use HTML5's adaptive quoting for HTML output from parse.js too. This now also
includes the body element, which gives us its dsr info and makes sure that the
content is fully parsed as the body even if re-parsed with a browser. Without
the body wrapper, we'd get the same issues as the ones described in
ebaef8dcb9.

Change-Id: I0fbd01171920efd1188b7be2a8e0ff638e7c24d9