Clone Wt2Html result before letting it become asynchronous.
The parser pipeline is handled asynchronously; ie the list of testTasks
is processed with async.waterfall() which defers work to process.nextTick().
The DOMPostProcessor is going to reset its parser as soon as it emits the
'document' event, which will clear parser.document.body.  We need to clone
the result before returning from the event handler to ensure that later
stages of the pipeline aren't affected by the imminent parser reset.

(This is necessary for domino and standards-compliant DOM implementations;
jsdom creates a new <body> when the parser is reset so the existing
<body> doesn't get stomped on.)

Change-Id: Ia372c5614a0b1aae846250a80374f2cabce6f202