Fix image parsing, which had been broken by the i18n fixes.
There was a bit of confusion in the image parsing code between
'canonical' mediawiki attribute names (which have img_ prefixes) the
unprefixed names used internally by the parser and renderer, and
localized forms used in the wikitext markup.

Added explicit comments documenting which is which, since they are
often the same when looking at English wikipedia.  We also made the
data stored in the optNames JSON attribute consistent (the key is the
"short form" of the canonical mediawiki attribute name).

Removed some "whitespace preservation" code, originally added in
88fc91a292428e873ea3fdd06f24a1ce59e7eeab.  The particular syntax
with whitespace surrounding the equals sign is not actually
supported by the PHP parser, and we stopped parsing it as well
with the i18n fixes in 6e4ef20e926f150250a4398b9e2bff9dfc176585.
So the code wasn't doing anything any more, and the thing it
was supposed to be doing it shouldn't have been doing in the first
place.  (We actually still pass the rt test case in 88fc91a, because
we preserve unparsed attributes literally.)

Tested with:

echo '[[Archivo:Nombre|miniatura|izquierda|enlace=foo|caption]]' |
 node tests/parse.js  --prefix=es --wt2wt

which is a handy way to keep clear the localized and nonlocalized
versions of these properties when stepping through the code.

Change-Id: Idac285b1537c2bc663178b032f5bdfe6d13e7cdd