Revert "Honor language=xx option in parserTests.txt"
This reverts commit dd593786a262c93272ef21c5f3d170313e253e02

We can't fetch configs from wikipedias for parserTests as those expect very basic configs vs. the customized ones in wikipedias. We'll need to provide base configs corresponding to a vanilla wiki in the respective language instead.

This patch als caused parserTests to abort (probably with specific versions of the request library), so backing this out for now.

Change-Id: I87638a2514eb0a8e013e5b4ad2a36350711279a6