Fix new page creation with selective serialization enabled
There is no point in retrieving an old revision for selective serialization if
we did not get an oldid passed in. This patch switches to full serialization
in that case without attempting to fetch the page source.

Error handlign in the getPageSrc helper is also slightly improved by checking
for the revision member first.

Change-Id: I2344e2106303c751a850fa9d296975dd844015d5