Exterminate global usage in defines
We used to define a few constructors and some state on the global object to
make it available to the PEG tokenizer. We have since found a way to pass the
state into the PEG tokenizer, but a few constructors were still added to the
global object in the defines module. The predecessor patch
I8810da3a5ec287d7ae9078463ba01472a87a492e adds explicit local definitions in
most remaining places. This patch finally does the same for the tokenizer and
removes the global definitions altogether.

Change-Id: If0ada2ea04b680b90caeb4fa76c46ca272e78d43