Suppress 'function in loop' warnings, fixing one bug (in ParserFunctions).
According to the jshint documentation, the 'function created in loop'
warning is intended to catch bugs like:

    var nums = [];
    for (var i = 0; i < 10; i++) {
      nums[i] = function (j) {
        return i + j;
      };
    }

where the loop variable is captured by the interior function, and
unexpected behavior results.  And there was indeed one bug of this
sort in ParserFunctions.  But the warning is too broad, and continues
to flag functions in loops even after the possible bugs (loop
variables used inside the function) have been fixed.  Use:

  /* jshint loopfunc: true */

to suppress the warning where we've verified that Bad Things aren't
happening.

Change-Id: I66fe9f0e5d7094d6eacda2f720abf51aeba53300