mediawiki.ApiRequest: Fix exception "Object has no method 'removeListener'"
Usually methods are not detacheable. This is also the case with methods
of nodejs EventEmitter.

This code yanked the listeners from the internals and called them manually
but in doing so they lost 'this' context (which is expected to be an instance
of EventEmitter or something that inherits from it).

It worked by luck as nodejs (until 0.8) called removeListener from the once()
callback from a saved reference to 'this' but that changed in 0.10 so now it
no longer works and, per nodejs:

"the fact that [it worked] was purely by accident, an implementation detail [..]"
- https://github.com/joyent/node/commit/d1b4dcd6ac

Error:
> starting parsing of Main_Page
>
> events.js:174
>     this.removeListener(type, g);
>          ^
> TypeError: Object #<Object> has no method 'removeListener'
>     at g (events.js:174:10)
>     at processSome (Parsoid/js/lib/mediawiki.ApiRequest.js:91:17)
>     at process._tickCallback (node.js:415:13)

Bug: 45994
Change-Id: Ifd7800f1d2949c4e0b0ce186755e26dc65b2d506