JSHint cleanups: forward declarations, disambiguate variables.
Rearrange code where it's possible to ensure functions appear before uses.

In other places, add an explicit forward declaration.  This isn't strictly
necessary, but it technically ensures that behavior is the same across
JavaScript implementations (the spec allows different initialization behavior)
and it shuts jshint up.  Seems easy enough to make jshint happy.

One case where two variables named 'tsr0' appear in the same scope; renamed
them slightly to avoid the conflict.  (This lets jshint catch if I missed a
renaming; if I'd just moved the declaration to the top of the method it
would be possible to hide a scope error.)

Change-Id: I35699cdd5cee02883a9f35fc5de3954080e72e0c