Fix stack space crasher with large tables.
For a page which has lots of table lines (not necessarily belonging to the
same table) the tokenizer would get into the following recursive call tree:

  nested_block -> block -> block_lines -> block_line -> table_lines ->
            table_line -> table_data_tags -> table_data_tag -> nested_block

Instead of finishing the table_data_tag with the next table_line, the
'nested_block*' in the table_data_tag production attempted to snarf all
succeeding blocks of the file (pushing 8 additional stack frames for
every line of the file).

Change the nested_block* production to exclude table_lines, preventing
the harmful recursion.

Test case is:
http://ru.wikipedia.org/wiki/Чемпионат_России_по_хоккею_с_мячом_1999/2000

Change-Id: I83b9803cd7a0f6450d8a4137fb245b7d36019e01