Fix wt2wt handling of link prefixes.
DU.getAttributeShadowInfo() never returns an object with a 'string' property.
I think this was an old refactoring that missed a case.  Changing the
test to 'target.value' lets us use splitLinkContentString to handle
link trails and link prefixes.  Since we're handling link trails with
this case now, we don't need the old special case for them, so that
simplifies the code a bit.

Two more parser tests pass.  (As well as the new parser test added in
https://gerrit.wikimedia.org/r/61617 which hasn't been merged yet.)

Change-Id: I1fdf9e60bfacd75808463e972422ff2d15ffa52f