Update parserTests, taking into account upstream edit section changes.
In https://gerrit.wikimedia.org/r/49364 (bug 41729) the edit section links
were moved after the headings, and all the upstream parser tests updated to
match.

Update our hashes so we fetch the new version of parserTests.txt from
upstream, and also update the regexp in Util.normalizeHTML we use to
remove edit section links.

We now pass 2 additional wt2html tests, because the improved regexp
better matches (the whitespace associated with) headings and because
upstream moved around the whitespace associated with the edit section
link in such a way that Parsoid can effectively ignore it.  Ie:

<h1> <span class="mw-headline" id="heading">heading</span></h1>

which didn't match our editsection regexp because of the leading
whitespace, became:

<h1><span class="mw-headline" id="heading">heading</span> </h1>

which now normalizes correctly to:

<h1>heading</h1>

...eating up the trailing whitespace (which is there to separate the
heading from the edit link, which is suppressed in this case).

Change-Id: I73fed7551e6ded38a24eb936ad015fe1e57d036d