Bug 48477: Fix new wiki link serialization
b5beb51e started to modify the dsr value of passed-in HTML nodes with this
line:

dp.dsr = dp.dsr || [];

This broke modification detection in DOMUtils.getAttributeShadowInfo, which
relied on Object.keys(dp) being empty. This lead to new links being treated as
if they were unmodified, resulting in bug 48477:

echo '<a href="Foo" rel="mw:WikiLink">foo</a>' | node parse --html2wt
[[Foo]]

This patch changes the 'new content' detection in
DOMUtils.getAttributeShadowInfo to directly check for the presence of the
data-parsoid attribute, which is not affected by the dsr initialization
introduced in b5beb51e.

Bug: 48477
Change-Id: I4d47e69a87dd957f0a6059d935d2fdee9d21cdfd