Fix text / paragraph separator
The following test case would fail to insert the needed two newlines between
the plain text and the paragraph:

echo 'foo<p>bar</p>' | node parse --html2wt

The issue was that a body parent was suspected to end up on the same line as
the text node, in which case the extra newline is not needed to delimit the
following paragraph.

This patch simply excludes the body element from the block-level parent check.
This is likely not a perfect solution, but solves the most common issue for
now.

Change-Id: I63b53edd9f16e99a4e220bf07153e861710e10f1