Encode request URL for cache requests
Cache requests for titles with special chars like ? were failing as the ? was
interpreted as the begin of the query string by our web API. That caused
selective serialization and expansion reuse to fail on titles like
Who_Wants_to_Be_a_Millionaire%3F_(Australian_game_show).

This patch adds percent encoding for the title, which fixes the issue.

Change-Id: I1437446961f3d7c56a4bed42a1383ed76a10d81d