Further row back on perf optimization
Only test heuristic at start of table cell and -caption attribute production,
but not for each attribute. This basically means that anything is eaten in the
middle of attributes, but should still avoid most backtracking on content.

Parse times on el:Πορεία_των_κυπριακών_ομάδων_στα_κύπελλα_Ευρώπης are
actually improved by this, as the sanity check is executed less often.

Change-Id: I1e982da057d10370972491e5e9b3a1225e5cb79d